Replace tycho-specific launcher icon path with copying the icon during build
Currently the launcher icon path is specific for tycho;
when exporting the product using PDE the launcher icon is not replaced.

By setting the icon path as PDE expects it
and copying the icon to the location tycho expects it during the build
the product can be exported with PDE and tycho.

Signed-off-by: mudelta <davymeers@hotmail.com>